CoreRT updates for files moved to shared partition
