+++ src/System.Private.CoreLib/shared/System/Runtime/InteropServices/ComInterfaceType.cs
@@ -1,0 +1,14 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.InteropServices
+{
+    public enum ComInterfaceType
+    {
+        InterfaceIsDual = 0,
+        InterfaceIsIUnknown = 1,
+        InterfaceIsIDispatch = 2,
+        InterfaceIsIInspectable = 3,
+    }
+}
