+++ src/System.Private.CoreLib/shared/System/Runtime/InteropServices/LCIDConversionAttribute.cs
@@ -1,0 +1,17 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.InteropServices
+{
+    [AttributeUsage(AttributeTargets.Method, Inherited = false)]
+    public sealed class LCIDConversionAttribute : Attribute
+    {
+        public LCIDConversionAttribute(int lcid)
+        {
+            Value = lcid;
+        }
+
+        public int Value { get; }
+    }
+}
