+++ src/System.Private.CoreLib/src/System/Runtime/InteropServices/ComTypes/IEnumerable.cs
